requirements.txt
torch
torchvision
numpy==1.26.4
matplotlib==3.9.0
Pillow==10.4.0
